added status card and outlines to circles
